Split kas::theme module for implementation
